requirements.txt
# Web app using Flash + ElasticSearch
flask==2.3.2
wtforms==3.0.1
flask-wtf==1.1.1
Werkzeug~=2.3.7

elasticsearch~=8.9.0
opensearch-py~=2.3.2

# required by '.env' and '.flaskenv' files
python-dotenv~=1.0.0

# Utilities
#Pillow~=9.3.0
#tqdm~=4.65.0
#numpy~=1.24.1



# Eland for import NLP models
torch==2.1.0
open-clip-torch==2.23.0
sentence-transformers~=2.2.2
#multilingual-clip==1.0.10
#eland>=8,<9bv





# for reading the exif jpg data
#exif==1.5.0

# required only for RAW images conversion
# rawpy==0.17.3
# imageio==2.22.4

